Prevent crashes caused by taking pointers to values that later get moved
The WriteBuilder takes pointers of values (wv and wvc) that get passed into
wavpack. When the builder moves the values into the resulting WriteContext,
the pointers become invalid. The fix here is to allocate the values on the heap;
the moved Boxes then maintain the heap allocated values, ensuring the pointers
given to wavpack remain valid.